Move preview controller out of account_manager.rs
