+++ src/ui/profile/mod.rs
@@ -1,5 +1,7 @@
 pub mod picture;
 pub mod preview;
+mod profile_preview_controller;
 
 pub use picture::ProfilePic;
 pub use preview::ProfilePreview;
+pub use profile_preview_controller::SimpleProfilePreviewController;
